refactor(core): modularize bot structure with cogs
The bot structure was refactored to improve maintainability and scalability.

- Move commands, events to separate cogs modules.
- Create init files for better package management.
- Update main.py to load cogs dynamically.
