+++ cogs/models/command_model.py
