+++ cogs/movies/detailed_movie.py
