+++ cogs/movies/detailed_series.py
