+++ cogs/spotify/cover_image.py
